move Readiness and SessionMetrics to network/mod.rs
